PersistenceCassandraConfigProvider provides Persistence Cassandra config
